Finish sentence trimming, need code review
