+++ services/transcribe/output/MeetingMinutes.txt
@@ -1,0 +1,8 @@
+Meeting Minutes
+
+Speaker 3: Good morning Professor Austin early
+Speaker 2: morning James. I'm doing well, and you I'm great.
+Speaker 3: Thank you. This is my friend Emma. She's thinking about applying to this college. She has a few questions. Would you mind solely us about this process, please?
+Speaker 2: Hello Emma. It's presently 2. I'm more than happy to we speak with you. Please stop by my office next week
+Speaker 1: and Sophia to meet you faster. Thank you so much for helping
+Speaker 2: the message. Hopefully I will be able to answer over your question.
